requirements.txt
spade==4.1.2
python-dotenv==1.2.1
pandas==2.3.3
matplotlib==3.10.7
SQLAlchemy==2.0.40
loguru==0.7.2
pyjabber==0.3.0
